Fix the nav menu
move menu button to the navbar dir and fix a nav bar on top.
