+++ src/App.tsx
@@ -3,11 +3,13 @@ import "./App.css";
 import Home from "./components/Home/Home";
 import Navigation from "./components/Navigation/Navigation";
 import { motion, AnimatePresence } from "framer-motion";
+import NavBar from "./components/NavBar/NavBar";
 
 function App() {
   const [open, setOpen] = useState<boolean>(false);
   return (
     <>
+      <NavBar open={open} setOpen={setOpen} />
       {open ? (
         <AnimatePresence>
           <motion.div
@@ -17,12 +19,12 @@ function App() {
             exit={{ y: "-100vh" }}
             transition={{ delay: 0.5, duration: 0.7, type: "tween" }}
           >
-            <Navigation open={open} setOpen={setOpen} />
+            <Navigation />
           </motion.div>
         </AnimatePresence>
       ) : (
         <div className="app">
-          <Home open={open} setOpen={setOpen} />
+          <Home />
         </div>
       )}
     </>
