+++ src/components/NavBar/MenuButton/MenuButton.tsx
@@ -5,11 +5,6 @@ type props = {
   open?: boolean;
 };
 
-const buttonvariants = {
-  visible: { opacity: 1, transition: { delay: 0.9, duration: 1.5 } },
-  hidden: { opacity: 0, transition: { delay: 0.9, duration: 1.5 } },
-};
-
 function MenuButton({ open }: props) {
   const variant = open ? "opened" : "closed";
   const top = {
@@ -41,12 +36,7 @@ function MenuButton({ open }: props) {
     },
   };
   return (
-    <motion.div
-      className="menu-container"
-      initial="hidden"
-      animate="visible"
-      variants={buttonvariants}
-    >
+    <motion.div className="menu-container">
       <motion.div
         transition={{ type: "spring", stiffness: 260, damping: 20 }}
         variants={top}
