+++ src/components/Navigation/Navigation.css
@@ -4,20 +4,7 @@
     display: flex;
     flex-direction: column;
     background-color: #ff3838;
-}
-.navigation-nav{
-    width: 100vw;
-    height: auto;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    background-color: rgb(20, 21, 121);
-}
-
-.navigation-hamburger-menu {
-    width: auto;
-    height: auto;
-    margin: 10px 20px;
+    z-index: 1;
 }
 .navigation-body {
     width: 100vw;
@@ -25,5 +12,5 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    background-color: antiquewhite;
+    
 }
