+++ src/components/Navigation/Navigation.tsx
@@ -1,20 +1,8 @@
-import { Dispatch, SetStateAction } from "react";
 import "./Navigation.css";
-import MenuButton from "../Home/MenuButton/MenuButton";
 
-type navProps = {
-  open?: boolean;
-  setOpen?: Dispatch<SetStateAction<boolean>>;
-};
-
-function Navigation({ open, setOpen }: navProps) {
+function Navigation() {
   return (
     <div className="navigation">
-      <div className="navigation-nav">
-        <div className="navigation-hamburger-menu">
-          <MenuButton open={open} />
-        </div>
-      </div>
       <div className="navigation-body"></div>
     </div>
   );
